Moved the witness client to be with the witness code
This will be reused by the sumdb feeder I'm in the process of writing.
